チェック時returnを入れた

+++ src/App.jsx
@@ -199,7 +199,8 @@ function App() {
               type='button'
               className='create-btn'
               onClick={()=>{
-                confirm("既存のデータがある場合上書きされますがよろしいですか？")
+                const ok = confirm("既存のデータがある場合上書きされますがよろしいですか？")
+                if(!ok) return
                 const obj = JSON.stringify({score:score})
                 localStorage.setItem("score",obj)
                 alert("ローカルストレージに譜面を保存しました。")
